+++ src/modules/providers/components/providers.tsx
@@ -1,0 +1,12 @@
+import { PropsWithChildren } from "react"
+import { ThemeProvider } from "./theme"
+import { headers } from "next/headers"
+import { TRPCReactProvider } from "./trpc-react"
+
+export const Providers = ({ children }: PropsWithChildren) => {
+    return (
+        <TRPCReactProvider headers={headers()}>
+            <ThemeProvider attribute="class">{children}</ThemeProvider>
+        </TRPCReactProvider>
+    )
+}
